% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_geom_text}
\alias{add_geom_text}
\title{Add Layer for Geometric Objects}
\usage{
add_geom_text(data_, active_entity, colors)
}
\arguments{
\item{data_}{Data frame}

\item{active_entity}{String. Currently active entity (variables vs
individuals)}

\item{color}{Hexadecimal. The color of the text}
}
\description{
This utility adds a text geometric object, conditional to which entity is
currently active (variables vs individuals).
We only want labels when plotting variables.
}
