% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBusyMessage.R
\name{addBusyMessage}
\alias{addBusyMessage}
\title{Module for Adding 'shiny-busy' Messages}
\usage{
addBusyMessage(ns)
}
\arguments{
\item{id}{Modules'ID}
}
\description{
This module improves User Experience by adding "Loading..." messages when
shiny gets busy making computations.
}
