% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{aling_plot_title}
\alias{aling_plot_title}
\title{Align plot title}
\usage{
aling_plot_title(position)
}
\arguments{
\item{position}{Admitted values are:
\itemize {
\item "left"
\item "center"
\item "right}}
}
\description{
Align plot title
}
