% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_style.R
\name{apply_custom_css}
\alias{apply_custom_css}
\title{Apply CSS Styles}
\usage{
apply_custom_css()
}
\value{
Tag List, each one containing a style tag applied to the head of the
html
}
\description{
This function applies custom defined css style.
}
