% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply Filter to MCA object}
\usage{
apply_filter(MCA_list, filter, value, dims)
}
\arguments{
\item{MCA_list}{List containing output from MCA.}

\item{filter}{Filter type. It can be "variables" or "individuals", depending on
which plot is currently active}

\item{value}{num. Value of the filter to be applied}

\item{dims}{array informing about currently active dimensions

If \code{filter = 'cos2'}, only rows with cos2 greater than or equal to
\code{value} in at least one of plotted dimensions will be kept.
If \code{filter = 'contrib'}, only the top \code{value} observations with
best contribution in at least one of the dimensions being displayed. For example,
if \code{value = 5} and \code{filter = 'contrib'}, an observation won't be
displayed if it does not rank among the top 5 contributions in any of
displayed dimensions.}
}
\value{
Data frame with only those observations from \code{DF} that pass the
filtering criteria, and the columns corresponding to current active dimensions
}
\description{
Apply Filter to MCA object
}
