% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAControlModule.R
\name{controlMCA_server}
\alias{controlMCA_server}
\title{Server-side processing}
\usage{
controlMCA_server(id, display_DF, displayed_vars, active_entity, active_dims)
}
\arguments{
\item{id}{The module's id}

\item{display_DF}{Reactive value, containing the data frame currently displayed
in the import section}

\item{displayed_vars}{Reactive value, containing the names of the variables
currently displayed in the import section}

\item{active_entity}{Reactive value, returning the active entity (individual,
variable)}

\item{active_dims}{Reactive value, returning currently selected dimensions}
}
\description{
Server-side processing
}
