% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAControlModule.R
\name{controlMCA_UI}
\alias{controlMCA_UI}
\title{Module for MCA Analysis Control}
\usage{
controlMCA_UI(id)
}
\arguments{
\item{id}{The module's id}
}
\value{
\code{shiny.tag.list} object containing \code{shiny} UI components
}
\description{
Module for MCA Analysis Control
}
