% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cos2_filter}
\alias{cos2_filter}
\title{Return TRUE if cos2 is greater than or equal to the filter value,
for at least one of the two dimensions, FALSE otherwise.}
\usage{
cos2_filter(DF, value)
}
\description{
Return TRUE if cos2 is greater than or equal to the filter value,
for at least one of the two dimensions, FALSE otherwise.
}
