% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_eigen_ggplot}
\alias{create_eigen_ggplot}
\title{Create Ggplot Object for Eigenvalues}
\usage{
create_eigen_ggplot(eigenvalues)
}
\arguments{
\item{eigenvalues}{Eigenvalues of current MCA results. Eigenvalues can be
expressed either as absolute values, or as cumulative percentage of
explained variance}
}
\description{
Create Ggplot Object for Eigenvalues
}
