% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_entities_ggplot}
\alias{create_entities_ggplot}
\title{Create Ggplot Object for Entities (Variables or Individuals)}
\usage{
create_entities_ggplot(DF, active_entity, active_dims, order_by)
}
\arguments{
\item{DF}{Data frame. The data to pass to ggplot}

\item{active_entity}{String. Currently active entity (variables vs
individuals)}

\item{active_dims}{Array of length 2. Currently active dimensions}

\item{order_by}{String. Must to be one of currently active dimensions.
Map the color of plotted points by cos2 of selected dimension}

\item{active_qualiquanti}{Reactive value, returning the array of currently
active qualitative and quantitative variables. Active variables are obtained
by taking the input given by the user and filtering out the variables whose
data types is not coherent with the category of the input (quantitative or
qualitative).}
}
\description{
Create Ggplot Object for Entities (Variables or Individuals)
}
