% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayImportModule.R
\name{displayImport_server}
\alias{displayImport_server}
\title{Server-side Processing for Displaying the Data Frame}
\usage{
displayImport_server(
  id,
  imported,
  selected,
  include_bttn,
  clear_bttn,
  display_DF,
  displayed_vars
)
}
\arguments{
\item{id}{Module's ID for namespacing}

\item{imported}{Reactive resulting from \code{importDataModule}, containing,
among other thing, the data frame as originally imported.}

\item{selected}{Reactive returned by \code{selectVarsModule}, containing the
array of variables the user wants to be displayed.}

\item{include_bttn}{Reactive value set upstream to trigger data frame displaying}

\item{clear_bttn}{Reactive value triggered upstream for displaying DF without
selected variables}

\item{display_DF}{Reactive values returning the data frame to be displayed,
updated with current selected variables and applied transformations}

\item{displayed_vars}{Reactive value returning currently displayed variables}
}
\description{
Server-side Processing for Displaying the Data Frame
}
