% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{font_family}
\alias{font_family}
\title{Return a thema function for customizing font}
\usage{
font_family(font_family, ...)
}
\arguments{
\item{font_family}{Font-family to apply}
}
\description{
Return a thema function for customizing font
}
