% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_best_set_by_contrib}
\alias{get_best_set_by_contrib}
\title{Get Best Set by Contribution}
\usage{
get_best_set_by_contrib(DF, contribs, value)
}
\arguments{
\item{DF}{Data frame}

\item{contribs}{Data frame containing contributions to active dimensions}

\item{value}{int. The value of the contribution filter}
}
\value{
Data frame with only those observations from \code{DF} that pass the
filtering criteria
}
\description{
This function is used for filtering observation in a data frame according to
their contribution to dimensions
}
