% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_descriptive_binaries}
\alias{get_descriptive_binaries}
\title{get_descriptive_binaries}
\usage{
get_descriptive_binaries(DF)
}
\arguments{
\item{DF}{A data frame}
}
\description{
Convert all 1/0 columns of the input data frame to Yes/No columns
}
