% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_eigens}
\alias{get_eigens}
\title{Utility function for getting eigenvalues of MCA factors (MCA is performed
using FactoMineR package)}
\usage{
get_eigens(MCA, cumulative_percent)
}
\arguments{
\item{MCA}{MCA object}

\item{cumulative_percent}{Logical. It True, extract the cumulative percentages
of explained variance, starting from the first MCA factor}
}
\value{
numerical array, with length equal to the number of MCA factor,
containing the eigenvalues associated with each of those factors, or the cumu-
lative percentages of explained variance if \code{cumulative_percent} is TRUE
}
\description{
Utility function for getting eigenvalues of MCA factors (MCA is performed
using FactoMineR package)
}
