% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_filter_value}
\alias{get_filter_value}
\title{Get Filter Value}
\usage{
get_filter_value(filter, cos2, contrib, active_DF)
}
\arguments{
\item{filter}{Name of the filter to be applied}

\item{cos2}{User-set value of cos2 filter}

\item{contrib}{User-set value of contrib filter}

\item{active_DF}{Currently active data frame}
}
\description{
this function returns the current value of the filter (be it cos2 or contrib)
}
