% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_match_index}
\alias{get_match_index}
\title{Find which labeled_value match to a certain row name.}
\usage{
get_match_index(row_name, labeled_values)
}
\description{
Find which labeled_value match to a certain row name.
}
