% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_mutated_DF}
\alias{get_mutated_DF}
\title{Get input DF mutated with input variables}
\usage{
get_mutated_DF(DF, transform_out, selected)
}
\arguments{
\item{DF}{A data frame}

\item{transform_out}{Data frame containing the result of the previous
transformation, to be added to \code{DF}}

\item{selected}{Character array containing which variables were included
in the logical operation}
}
\value{
a data frame, mutated with arrays resulting from previous transforma-
tion step
}
\description{
Get input DF mutated with input variables
}
