% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_validated_DF}
\alias{get_validated_DF}
\title{Get Validated Data Frame}
\usage{
get_validated_DF(DF)
}
\arguments{
\item{DF}{A data frame}
}
\value{
A data frame
}
\description{
Take a data frame, and remove all columns belonging to a predefined set of
data types.
}
