% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_var_output}
\alias{get_var_output}
\title{Utility for getting MCA output for active variables}
\usage{
get_var_output(MCA, labeled_values)
}
\arguments{
\item{MCA}{MCA object}

\item{labeled_values}{Array, with values labeled by category}
}
\description{
Utility for getting MCA output for active variables
}
