% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataModule.R
\name{importData_UI}
\alias{importData_UI}
\title{Module for Importing Data}
\usage{
importData_UI(id)
}
\arguments{
\item{id}{Module's ID for namespacing}
}
\value{
\code{\link[shiny]{tagList}} that can be passed to a shinyUI function
}
\description{
Provides interface for uploading pairs of csv and xml files.
}
\details{
After a pair of files is uploaded, the user can clicking on the \code{import}
button to display the list of all variable names extracted by imported data.
This documentation refers to the UI-side function of the module. For the documentation
about the server-side function, got to \code{\link{importData_server}}
}
