% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_binary}
\alias{is_binary}
\title{Utility returning TRUE for binary data, FALSE for non-binary data}
\usage{
is_binary(array_)
}
\arguments{
\item{Array}{}
}
\value{
Logical value
}
\description{
Utility returning TRUE for binary data, FALSE for non-binary data
}
