% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{justify_plot_title}
\alias{justify_plot_title}
\title{Justify plot title}
\usage{
justify_plot_title(position)
}
\arguments{
\item{position}{Admitted values are:
\itemize {
\item "left"
\item "center"
\item "right}}
}
\description{
Justify plot title
}
