% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mental_health_tech}
\alias{mental_health_tech}
\title{Mental Health in Tech Survey
Survey on Mental Health in the Tech Workplace in 2014}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1259 rows and 24 columns.
}
\source{
\url{https://www.kaggle.com/osmi/mental-health-in-tech-survey}
}
\usage{
mental_health_tech
}
\description{
Mental Health in Tech Survey
Survey on Mental Health in the Tech Workplace in 2014
}
\keyword{datasets}
