% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_qualiquanti}
\alias{merge_qualiquanti}
\title{Extend output of MCA}
\usage{
merge_qualiquanti(out, qualiquanti_out)
}
\arguments{
\item{out}{Dataframe, with output of entities}

\item{qualiquanti_out}{list, with output of qualitative and quantitative variables}
}
\description{
Extend output of MCA
}
