% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{minimal_variance_filter}
\alias{minimal_variance_filter}
\title{Apply Variance Filter to Variables}
\usage{
minimal_variance_filter(DF, threshold)
}
\arguments{
\item{DF}{A data frame}

\item{threshold}{num. Variables whose variance is lesser than
\code{threshold} is removed from \code{DF}}
}
\value{
\code{DF}, with variables within the given variance threshold removed
}
\description{
This function takes a data frame, and transforms it by keeping only those
binary variables whose variance is greater than a given threshold.
Non-binary variables are kept as they are
}
