% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCAModule.R
\name{plotMCA_server}
\alias{plotMCA_server}
\title{Server-side processing}
\usage{
plotMCA_server(id, plot_DF, eigenvalues, active_entity, active_dims)
}
\arguments{
\item{id}{Modules's id}

\item{plot_DF}{Reactive value containing the dataframe to be plotted.}

\item{eigenvalues}{Reactive value containing the eigenvalues of MCA factors.}

\item{active_entity}{Reactive value, returning the active entity (individual,
variable)}

\item{active_dims}{Reactive value, returning currently selected dimensions}
}
\description{
Server-side processing
}
