% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils.R
\name{read_inst}
\alias{read_inst}
\title{Read a RDS file from the inst/ directory}
\usage{
read_inst(file_name)
}
\arguments{
\item{file_name}{Name of the file to be read}
}
\description{
Read a RDS file from the inst/ directory
}
