% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_rows_with_NAs}
\alias{remove_rows_with_NAs}
\title{remove_rows_with_NAs}
\usage{
remove_rows_with_NAs(DF)
}
\arguments{
\item{DF}{A data frame}
}
\value{
A data frame
}
\description{
This function takes a data frame as input, and removes all rows having at
at least one missing value.
}
