% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils.R
\name{save_inst}
\alias{save_inst}
\title{Save a RDS file into the inst/ directory}
\usage{
save_inst(obj, file_name)
}
\arguments{
\item{file_name}{Name of the file to be read}
}
\description{
Save a RDS file into the inst/ directory
}
