% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVarsModule.R
\name{selectVars_server}
\alias{selectVars_server}
\title{Server-side Processing for Returning User-selected Variables}
\usage{
selectVars_server(id, imported, selected, include_bttn, clear_bttn)
}
\arguments{
\item{id}{Module's ID}

\item{imported}{reactive value containing up-to-date data imported by the user
(this is shared among modules)}

\item{selected}{Reactive returned by \code{selectVarsModule}, containing the
array of variables the user wants to be displayed.}

\item{include_bttn}{Reactive value used to trigger downstream dependencies}

\item{clear_bttn}{Reactive value used to trigger downstream dependencies}
}
\description{
Server-side Processing for Returning User-selected Variables
}
