% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils.R
\name{setInputs_test_plot_DF}
\alias{setInputs_test_plot_DF}
\title{Specific-purpose Utility for Setting Test Inputs}
\usage{
setInputs_test_plot_DF(a_p, f, cos2_value, contrib_value, session)
}
\arguments{
\item{a_p}{String. The currently active plot}

\item{f}{String. Filter to be applied}

\item{cos2_value}{Numerical. Value assigned to cos2 filter}

\item{contrib_value}{Numerical. Value assigned to contrib filter}

\item{session.}{Shiny session object}
}
\description{
Specific-purpose Utility for Setting Test Inputs
}
