% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{transform_}
\alias{transform_}
\title{Apply Logical Transformation to Selected Variables}
\usage{
transform_(DF, selected, operation, new_var_name, discard_selected)
}
\arguments{
\item{DF}{A data frame}

\item{selected}{Character array containign which variables are to be included
in the logical operation}

\item{operation}{String, specifying which logical operation is to be applied
to the set of selected variables. Admitted values are:
\describe {
\item{AND}
\item{OR}
\itemè{NOT --> Only works for one variables set}
}}

\item{new_var_name}{String, naming the newly created variable}

\item{discard_selected}{Logical. Should variables from the transformation pool
be discarded?}
}
\value{
dataframe, containing the results of the logical operation
}
\description{
This function is designed for applying logical operation to the set of sele-
cted variables, extracted from a given dataframe.
}
