% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformNumericalModule.R
\name{transformNumerical_UI}
\alias{transformNumerical_UI}
\title{Module for Transforming Numerical Data}
\usage{
transformNumerical_UI(id)
}
\arguments{
\item{id}{Module's ID for namespacing}
}
\value{
\code{\link[shiny]{tagList}} that can be passed to a shinyUI function
}
\description{
This module can be used for creating synthetic variables.
UI function
}
