% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformSurveyModule.R
\name{transformSurvey_server}
\alias{transformSurvey_server}
\title{Server-side Processing for Applying Transformations to Current Data Frame}
\usage{
transformSurvey_server(id, display_DF, displayed_vars)
}
\arguments{
\item{id}{Module's ID for namespacing}

\item{display_DF}{Reactive value returning the data frame to be displayed,
updated with current selected variables and applied transformations}

\item{displayed_vars}{Reactive value returning currently displayed variables}
}
\description{
Server-side Processing for Applying Transformations to Current Data Frame
}
