% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformSurveyModule.R
\name{transformSurvey_UI}
\alias{transformSurvey_UI}
\title{Module for Transforming Survey Data}
\usage{
transformSurvey_UI(id)
}
\arguments{
\item{id}{Module's ID for namespacing}
}
\value{
\code{\link[shiny]{tagList}} that can be passed to a shinyUI function
}
\description{
This module can be used for creating synthetic variables.
UI function
}
