% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_factor_analysis.R
\name{UI_factor_analysis}
\alias{UI_factor_analysis}
\title{UI for Importing and Crafting Data}
\usage{
UI_factor_analysis()
}
\value{
A UI definition that can be passed to the shinyUI function
}
\description{
UI for Importing and Crafting Data
}
