% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_import.R
\name{UI_import}
\alias{UI_import}
\title{UI for Importing and Crafting Data}
\usage{
UI_import(
  import_id = "import",
  select_id = "select",
  display_id = "display_import"
)
}
\arguments{
\item{import_id}{ID of \code{importDataModule}}

\item{select_id}{ID of \code{selectVarsModule}}

\item{display_id}{ID of \code{displayDataModule} module}
}
\value{
A UI definition that can be passed to the shinyUI function
}
\description{
UI for Importing and Crafting Data
}
