% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_transform.R
\name{UI_transform}
\alias{UI_transform}
\title{UI for Transforming the Data}
\usage{
UI_transform(transform_id = "transform", display_id = "display_transform")
}
\arguments{
\item{transform_id}{ID of \code{transformDataModule} module}
}
\value{
A UI definition that can be passed to the shinyUI function
}
\description{
UI for Transforming the Data
}
