% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_transform_survey.R
\name{UI_transform_survey}
\alias{UI_transform_survey}
\title{UI for Transforming Survey Data}
\usage{
UI_transform_survey(transform_survey_id = "transform_survey")
}
\value{
A UI definition that can be passed to the shinyUI function
}
\description{
UI for Transforming Survey Data
}
