% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{variance_binary}
\alias{variance_binary}
\title{Compute Variance of Binary Variable}
\usage{
variance_binary(bin_array)
}
\arguments{
\item{bin_array}{Array of binary data (1/0, or TRUE/FALSE)}
}
\description{
This function returns the variance of a binary variable, compute as
\code{p(1-p)}, where \code{p} is the proportion of ones in the data (this
is the variance of a bernoullian distribution)
}
