% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{which_binary}
\alias{which_binary}
\title{Utility for detecting binary columns of a data frame}
\usage{
which_binary(DF)
}
\arguments{
\item{DF}{A data frame}
}
\value{
Logical array of TRUE and FALSE
}
\description{
Utility for detecting binary columns of a data frame
}
