% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataModule.R
\name{importData_server}
\alias{importData_server}
\title{Server-side Processing for Importing Data}
\usage{
importData_server(id, imported)
}
\arguments{
\item{id}{Module's ID for namespacing}

\item{imported}{reactive value containing up-to-date data imported by the user
(this is shared among modules)}
}
\description{
Server-side Processing for Importing Data
}
